third_party/WebKit/LayoutTests/fast/html/imports/import-dynamic-shared.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<link id="harness" rel="import" href="resources/async-child.html">
</head>
<body>
<script>
description("This test ensures that de-duped import can be dynamically added.");

window.jsTestIsAsync = true;

var linkEl = document.createElement("link");
linkEl.setAttribute("rel", "import");
linkEl.setAttribute("href", "resources/async-child.html");
document.head.appendChild(linkEl);
linkEl.onload = function(evt) {
    shouldBe("linkEl.import", "harness.import");
    finishJSTest();
};

</script>
</body>
</html>
