third_party/WebKit/LayoutTests/fast/html/imports/import-dynamic-shared-expected.txt

This test ensures that de-duped import can be dynamically added.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS linkEl.import is harness.import
PASS successfullyParsed is true

TEST COMPLETE
